Extract details/preview tabs sections to their own files, share prop types
